+++ RecastCLI/Source/MeshLoaderObj.cpp
@@ -228,7 +228,7 @@ bool rcMeshLoaderObj::load(const std::string& fileName)
 		n[0] = e0[1]*e1[2] - e0[2]*e1[1];
 		n[1] = e0[2]*e1[0] - e0[0]*e1[2];
 		n[2] = e0[0]*e1[1] - e0[1]*e1[0];
-		float d = sqrtf(n[0]*n[0] + n[1]*n[1] + n[2]*n[2]);
+		float d = std::sqrt(n[0]*n[0] + n[1]*n[1] + n[2]*n[2]);
 		if (d > 0)
 		{
 			d = 1.0f/d;
